+++ sphae.sh
@@ -14,7 +14,7 @@
 #sphae install
 sphae run --input tests/data/illumina-subset --threads 64 -k --use-singularity --sdm apptainer --use-conda
 sphae run --input tests/data/nanopore-subset --sequencing longread --threads 64 -k --use-singularity --sdm apptainer --use-conda
-#sphae run --input tests/data/nanopore-subset --sequencing longread --threads 64 -k --no_medaka  --use-singularity --sdm apptainer --use-conda
-#sphae annotate --genome /home/nala0006/scratch/mycobac --output mycob_test --threads 64
+sphae run --input tests/data/nanopore-subset --sequencing longread --threads 64 -k --no_medaka  --use-singularity --sdm apptainer --use-conda
+sphae annotate --genome tests/data/genome --threads 64 -k --use-singularity --sdm apptainer --use-conda
 
 #sphae run --input tests/data/illumina-subset/ --threads 64 -k --db_dir tests/db --conda-create-envs-only